target_device/SenseHatOnRaspberryPi/modules/SenseHatDisplay/requirements.txt
azure-iot-device~=2.0.0
RTIMULib
sense-hat